38-FE: Move shared form components to separate folder
