+++ Okane.Client/src/shared/components/form/FormSelect.spec.ts
@@ -2,7 +2,7 @@
 import { expect } from 'vitest'
 
 // Internal
-import FormSelect, { type FormSelectProps } from '@shared/components/FormSelect.vue'
+import FormSelect, { type FormSelectProps } from '@shared/components/form/FormSelect.vue'
 
 import * as formUtils from '@shared/utils/form.utils'
